ref(trace-view): Move shared tree toggle component
Ripped from #25264
