+++ static/app/views/performance/traceDetails/styles.tsx
@@ -9,7 +9,7 @@ import {SpanRow} from 'app/components/events/interfaces/spans/styles';
 import {Panel} from 'app/components/panels';
 import Pills from 'app/components/pills';
 import SearchBar from 'app/components/searchBar';
-import {IconChevron, IconFire} from 'app/icons';
+import {IconFire} from 'app/icons';
 import overflowEllipsis from 'app/styles/overflowEllipsis';
 import space from 'app/styles/space';
 import {Organization} from 'app/types';
@@ -20,7 +20,6 @@ import {Theme} from 'app/utils/theme';
 import {transactionSummaryRouteWithQuery} from 'app/views/performance/transactionSummary/utils';
 
 export {
-  ConnectorBar,
   DividerLine,
   DividerLineGhostContainer,
   DurationPill,
@@ -29,9 +28,6 @@ export {
   SpanBarTitleContainer as TransactionBarTitleContainer,
   SpanRowCell as TransactionRowCell,
   SpanRowCellContainer as TransactionRowCellContainer,
-  SpanTreeConnector as TransactionTreeConnector,
-  SpanTreeToggler as TransactionTreeToggle,
-  SpanTreeTogglerContainer as TransactionTreeToggleContainer,
 } from 'app/components/events/interfaces/spans/spanBar';
 
 export {
@@ -86,11 +82,6 @@ export const TraceViewContainer = styled('div')`
 
 export const StyledPanel = styled(Panel)`
   overflow: hidden;
-`;
-
-export const StyledIconChevron = styled(IconChevron)`
-  width: 7px;
-  margin-left: ${space(0.25)};
 `;
 
 export const TransactionRow = styled(SpanRow)<{cursor: 'pointer' | 'default'}>`
